LPD-69168 move classes to right package
